añadido la opcion de ir a signup en caso de no tener cuenta y querer crearse una y ir a main en caso de no querer iniciar sesion o cuando incia sesion y le da al boton
