--- app/src/main/java/com/canhub/canhub/activity_login.java
@@ -1,39 +1,0 @@
-package com.canhub.canhub;
-
-import android.os.Bundle;
-import android.widget.ImageView;
-
-import androidx.activity.EdgeToEdge;
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.core.graphics.Insets;
-import androidx.core.view.ViewCompat;
-import androidx.core.view.WindowInsetsCompat;
-
-import com.bumptech.glide.Glide;
-import com.bumptech.glide.load.resource.drawable.DrawableTransitionOptions;
-
-public class activity_login extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        EdgeToEdge.enable(this);
-        setContentView(R.layout.activity_login);
-
-//        ImageView canhub = findViewById(R.id.logo);//girl del id del xml
-//
-//        Glide.with(this)
-//                .load("https://images.unsplash.com/photo-1547355253-ff0740f6e8c1?q=80&w=1854&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D")
-//                //.load(R.drawable.girl)//girl.png
-//                .transition(DrawableTransitionOptions.withCrossFade(1000))
-//                .centerCrop()
-//                //  .circleCrop()
-//                //  .placeholder(new ColorDrawable(this.getResources().getColor(R.color.amarillo)))
-//                .into(canhub);
-        ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main), (v, insets) -> {
-            Insets systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars());
-            v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom);
-            return insets;
-        });
-    }
-}
