+++ app/src/main/res/values/colors.xml
@@ -1,15 +1,13 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <color name="negro">#FF000000</color>
-    <color name="blanco">#FFFFFFFF</color>
-    <color name="azul_claro">#6ECAEE</color>
-    <color name="azul">#048ABF</color>
-    <color name="azul_oscuro">#0477BF</color>
-    <color name="azulito">#0477BE</color>
+    <color name="black">#FF000000</color>
+    <color name="white">#FFFFFFFF</color>
+    <color name="azulclaro">#6ECAEE</color>
+    <color name="azul">#0477BF</color>
+    <color name="azulverde">#048ABF</color>
     <color name="amarillo">#F2C166</color>
     <color name="naranja">#F28705</color>
-    <color name="gris_claro">#F7F8F9</color>
-    <color name="gris">#8391A1</color>
+    <color name="grisclaro">#F2F2F2</color>
 
 
 </resources>
