--- app/src/main/res/values/style.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <style name="buttonflat">
-        <item name="android:layout_width">0dp</item>
-        <item name="android:layout_height">60dp</item>
-        <item name="android:backgroundTint">@color/azulito</item>
-        <item name="android:textColor">@color/blanco</item>
-        <item name="cornerRadius">10dp</item>
-    </style>
-</resources>
